pyproject.toml
[tool.poetry]
name = "project-fenaverat"
version = "0.1.0"
description = ""
authors = ["lxRbckl <lxrbckl@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
dash = "^2.17.1"
dash-bootstrap-components = "^1.6.0"
lxrbckl = "^3.3.3"
pygithub = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
